+++ Goose/sql/pets.sql
@@ -13,8 +13,8 @@ CREATE TABLE pets (
   experience_sold BIGINT DEFAULT 0 NOT NULL,
   aggro_range SMALLINT DEFAULT 0 NOT NULL,
   attack_range SMALLINT DEFAULT 0 NOT NULL,
-  attack_speed DECIMAL(9,4) DEFAULT 2 NOT NULL,
-  move_speed DECIMAL(9,4) DEFAULT 2 NOT NULL,
+  attack_speed REAL DEFAULT 2 NOT NULL,
+  move_speed REAL DEFAULT 2 NOT NULL,
   
   pet_hp INT DEFAULT 0 NOT NULL,
   pet_mp INT DEFAULT 0 NOT NULL,
@@ -46,9 +46,9 @@ CREATE TABLE pets (
   equipped_items TEXT DEFAULT '0,*,0,*,0,*,0,*,0,*,0,*' NOT NULL,
   weapon_damage INT DEFAULT 1 NOT NULL,
   
-  hp_percent_regen DECIMAL(9,4) DEFAULT 0 NOT NULL,
+  hp_percent_regen REAL DEFAULT 0 NOT NULL,
   hp_static_regen INT DEFAULT 0 NOT NULL,
-  mp_percent_regen DECIMAL(9,4) DEFAULT 0 NOT NULL,
+  mp_percent_regen REAL DEFAULT 0 NOT NULL,
   mp_static_regen INT DEFAULT 0 NOT NULL
 );
